RobotConfig made and sendable Chooser added
In this commit I add the RobotConfig class, which is a renamed IODEf that performs a similar function. It allows
for switching between the demobot and competitionbot, nicknamed falseKorea and trueKorea, respectively. I added a
sendable chooser that allows the user to pick the correct robot configuration from the smart dashboard. This is then
used to set all of the correct variables in autoInit, teleopInit, and testInit.