Tweaking PID
In the DriveTrain class, I've added the maxSpeed_inPs variable
which holds the maximum speed the PID should control to. I've
also added gain variables in the constructor to hold the same
value across the PIC controllers. These values seem to work
well, it's not trying to rip itself apart, but there's probably
smarter values to go here.

Outside of this, I've removed the driveTrain.drive() call from
testPeriodic() since this is already in the teleop method. Also,
I've added the math for the distancePerPulse calculation in the
RobotConfig class.

Unfortunately, since this branch wasn't pushed to GitHub, I had
to work in Colman's area... :(

- Patrick