Deletes Non-empty Dirs and More Comments
I've added a little bit of extra code to guarentee that the caller will be in the
directory they started in if an error occurs. Non-empty directories are now also
removed from the higher levels as necessary. There are some more comments to explain a
bit more about what's going on.