Fixed Library Names
It appears that the library load issue encountered when I loaded
SimpleStuart was caused by the library names in the .sdf file. All
these libraries are named like "libgz_clock.so" when they should be
"libclock.so". I've manually changed the library names, and the
model executes in Gazebo. Unfortuntately, it doesn't move when
commanded from the code.

We need to:
- Change the way the .sdf is generated to use the correct libraries
- Determine why commanding the wheels does not move the robot

Also in this commit, a little frcsim launcher to set the path
correctly for this new directory and a new ignore pattern for tildy
backup files.