Updated Dynamics, Limits, and Multipliers
I've updated the .sdf file to include non-zero joint axis dynamics
and limits, as well as plugin multipliers. The values here match
the values from the EKIPS robot installed with Gazebo and frcsim.
With these values, I can spin the wheels at a reasonable speed, but
the robot doesn't go anywhere. It's as if there is no friction
between the wheels and the ground plane. There may be some additional
physics that need to be added to the .sdf file. I've tried adding
physics manually from the EKIPS .sdf, but I've left with an error.