First Initial Winch and Other Drive Improvements
This is the last set of changes from Suffield Shakedown. Colman
added an initial Winch class to test the climber on the robot. The
climb speed was supposed to be adjustable by the throttle on one of
the joysticks, but that never appeared to work. There is also now a
non-linear profile to the joystick output.