Fixed Auto Selector and Added Test Selector
Using the default option in auto causes the code to have a null pointer error.
A quick change to use an empty string instead of Null clears that up.

I was also switching back and forth a lot between LiveWindow and custom test
modes, so I added a similar selector to the DriverStation to help with this. :)